prod_app/requirements.txt
appengine-python-standard==1.1.4
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
frozendict==2.3.8
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mock==5.1.0
packaging==23.1
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
zipp==3.16.2
